+++ app/src/main/java/com/example/githubclient/domain/entities/RepoDTO.kt
@@ -3,5 +3,5 @@ package com.example.githubclient.domain.entities
 data class RepoDTO(
     val id: Int = 0,
     val name: String = "",
-    val userId: Int = 0
+    val userId: Long = 0
 )
